apps/backend/requirements.txt
# RAG_GOD Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
websockets==12.0

# LangGraph and LangChain
langgraph==0.0.20
langchain==0.1.0
langchain-core==0.1.10

# Data processing
numpy==1.26.2
pandas==2.1.4

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# Optional: Mem0 for memory persistence (uncomment when ready)
# mem0ai==0.1.0

# Optional: Polymarket CLOB client (uncomment when ready)
# py-clob-client==0.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.23.2
